@@ -1,18 +1,31 @@
 {
   "name": "fish3046/lime-light-crm-php",
-  "description": "Sticky.io CRM v1 & v2 PHP Client",
+  "description": "Sticky.io v1 & v2 PHP Client",
+  "type": "library",
+  "license": "MIT",
+  "homepage": "https://github.com/fish3046/lime-light-crm-php",
+  "keywords": [
+    "sticky.io", "limelight"
+  ],
   "require": {
-    "php": ">=7.4.0",
+    "php": ">=8.1",
     "ext-json": "*",
-    "guzzlehttp/guzzle": "^6.2|^7.0"
+    "guzzlehttp/guzzle": "^7.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9",
-    "mockery/mockery": "^1.2",
+    "caseyamcl/guzzle_retry_middleware": "^2.8",
     "fakerphp/faker": "^1.6",
+    "illuminate/support": "^8.12",
+    "mockery/mockery": "^1.2",
+    "phpunit/phpunit": "^9",
+    "psr/log": "^3.0",
+    "squizlabs/php_codesniffer": "^3.7",
     "vlucas/phpdotenv": "^5.0"
   },
-  "license": "MIT",
+  "suggest": {
+    "caseyamcl/guzzle_retry_middleware": "Allows automatic retry of requests",
+    "psr/log": "Logging of retry requests"
+  },
   "authors": [
     {
       "name": "Kevin Em",
@@ -21,6 +34,10 @@
     {
       "name": "Ramil Vildanov",
       "email": "ramil@almeydahost.com"
+    },
+    {
+      "name": "Karl Pierce",
+      "email": "kpierce@ratespecial.com"
     }
   ],
   "autoload": {
@@ -32,5 +49,20 @@
     "psr-4": {
       "KevinEm\\LimeLightCRM\\Tests\\": "tests/"
     }
+  },
+  "scripts": {
+    "test": "phpunit",
+    "check-style": "phpcs src tests",
+    "fix-style": "phpcbf src tests"
+  },
+  "extra": {
+    "laravel": {
+      "providers": [
+        "KevinEm\\LimeLightCRM\\Laravel\\LimeLightCRMServiceProvider"
+      ]
+    }
+  },
+  "config": {
+    "sort-packages": true
   }
 }
