@@ -2,6 +2,7 @@
 
 namespace KevinEm\LimeLightCRM\Services;
 
+use GuzzleHttp\Exception\GuzzleException;
 use KevinEm\LimeLightCRM\v1\LimeLightCRM;
 
 /**
@@ -11,22 +12,17 @@ use KevinEm\LimeLightCRM\v1\LimeLightCRM;
  */
 class ProspectService
 {
-    protected LimeLightCRM $v1Engine;
-    protected \KevinEm\LimeLightCRM\v2\LimeLightCRM $v2Engine;
-
-    public function __construct(LimeLightCRM $v1Engine, \KevinEm\LimeLightCRM\v2\LimeLightCRM $v2Engine)
+    public function __construct(protected LimeLightCRM $v1Engine, protected \KevinEm\LimeLightCRM\v2\LimeLightCRM $v2Engine)
     {
-        $this->v1Engine = $v1Engine;
-        $this->v2Engine = $v2Engine;
     }
 
     /**
-     * @param       $prospectData
+     * @param array $prospectData
      * @param array $customFields
      * @return int                  Prospect ID
-     * @throws \GuzzleHttp\Exception\GuzzleException
+     * @throws GuzzleException
      */
-    public function addProspectWithCustomFields($prospectData, array $customFields = []): int
+    public function addProspectWithCustomFields(array $prospectData, array $customFields = []): int
     {
         $resp = $this->v1Engine
             ->prospects()
