@@ -1,8 +1,6 @@
 <?php
 
-
 namespace KevinEm\LimeLightCRM\v1;
-
 
 /**
  * Class Credentials
@@ -10,7 +8,7 @@ namespace KevinEm\LimeLightCRM\v1;
  */
 class Credentials extends AbstractService
 {
-    public function validateCredentials()
+    public function validateCredentials(): Response
     {
         return $this->makeRequest('validate_credentials', []);
     }
